requirements.txt
# Dependências principais do Sistema de Triagem ODQ
msal>=1.20.0
requests>=2.28.0
psutil>=5.9.0
python-dotenv>=1.0.0

# Backend FastAPI
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Autenticação JWT
PyJWT>=2.8.0

# Processamento de arquivos
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0

# Email processing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Cache Redis (opcional)
redis>=5.0.0
